+++ database/migrations/2018_10_27_160855_create_tbl_kategori.php
@@ -1,0 +1,34 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class CreateTblKategori extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('tbl_kategori', function (Blueprint $table) {
+            $table->increments('id');
+            $table->string('kategori_kode', 10)->unique();
+            $table->string('kategori_nama');
+            $table->string('kategori_slug');
+            $table->timestamps();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('tbl_kategori');
+    }
+}