+++ resources/views/staff/dashboard/index.blade.php
@@ -2,12 +2,12 @@
 
 {{-- Content Header --}}
 @section('staff_content_title')
-    <h1 class="m-0 text-dark">Dashboard v3</h1>
+    <h1 class="m-0 text-dark">Dashboard</h1>
 @endsection
 @section('staff_content_breadcrumb')
     <ol class="breadcrumb float-sm-right">
-        <li class="breadcrumb-item"><a href="#">Home</a></li>
-        <li class="breadcrumb-item active">Dashboard v3</li>
+        <li class="breadcrumb-item"><a href="{{ url('/staff') }}">Home</a></li>
+        <li class="breadcrumb-item active">Dashboard</li>
     </ol>
 @endsection{{-- Content Header --}}
 
@@ -20,4 +20,12 @@
             <a onclick="showSwal()" class="btn btn-primary">Show SweetAlert2</a>
         </div>
     </div>
+@endsection
+
+@section('inline_js')
+<script>
+    $(document).ready(function(){
+        $("#mn-dashboard").addClass('active');
+    });
+</script>
 @endsection